requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==1.7.1
sqlparse==0.5.3
tzdata==2025.1
